Virtual-Citizen-Assistant/requirements.txt
semantic-kernel==1.37.0
azure-search-documents==11.5.3
azure-identity==1.19.0
azure-ai-textanalytics==5.3.0
openai>=1.98.0
flask==3.1.0
python-dotenv==1.0.1
requests>=2.32.0
aiohttp>=3.11.0